+++ packages/aws-amplify-react/src/Auth/AuthPiece.jsx
@@ -11,7 +11,7 @@
  * and limitations under the License.
  */
 
-import React, { Component } from 'react';
+import { Component } from 'react';
 import { ConsoleLogger as Logger } from '@aws-amplify/core';
 
 import AmplifyTheme from '../Amplify-UI/Amplify-UI-Theme';
@@ -81,6 +81,7 @@ export default class AuthPiece extends Component {
     render() {
         if (!this._validAuthStates.includes(this.props.authState)) {
             this._isHidden = true;
+            this.inputs = {};
             return null;
         }