+++ packages/aws-amplify-react/src/Auth/TOTPSetup.jsx
@@ -11,7 +11,7 @@
  * and limitations under the License.
  */
 
-import React, { Component } from 'react';
+import * as React from 'react';
 import { JS, ConsoleLogger as Logger } from '@aws-amplify/core';
 import Auth from '@aws-amplify/auth';
 import AuthPiece from './AuthPiece';
@@ -37,15 +37,15 @@ export default class TOTPSetup extends AuthPiece {
                 if (!JS.isEmpty(data.verified)) {
                     this.changeState('signedIn', user);
                 } else {
-                    user = Object.assign(user, data);
-                    this.changeState('verifyContact', user);
+                    const newUser = Object.assign(user, data);
+                    this.changeState('verifyContact', newUser);
                 }
             });
     }
 
     onTOTPEvent(event, data, user) {
         logger.debug('on totp event', event, data);
-        //const user = this.props.authData;
+        // const user = this.props.authData;
         if (event === 'Setup TOTP') {
             if (data === 'SUCCESS') {
                 this.checkContact(user);
@@ -59,6 +59,6 @@ export default class TOTPSetup extends AuthPiece {
 
         return (
             <TOTPSetupComp {...this.props} onTOTPEvent={this.onTOTPEvent} />
-        )
+        );
     }
 }