+++ packages/aws-amplify-react/src/Storage/S3Album.js
@@ -11,7 +11,8 @@
  * and limitations under the License.
  */
 
-import React, { Component } from 'react';
+import * as React from 'react';
+import { Component } from 'react';
 
 import { ClientDevice, JS, ConsoleLogger as Logger, Hub } from '@aws-amplify/core';
 import Storage from '@aws-amplify/storage';
@@ -35,7 +36,7 @@ export default class S3Album extends Component {
 
         const theme = this.props.theme || AmplifyTheme;
         this.state = {
-            theme: theme,
+            theme,
             items: [],
             ts: new Date().getTime()
         };
@@ -53,7 +54,7 @@ export default class S3Album extends Component {
             if (callback_type === 'string') {
                 key = fileToKey;
             } else if (callback_type === 'function') {
-                key = fileToKey({ name: name, size: size, type: type });
+                key = fileToKey({ name, size, type });
             } else {
                 key = encodeURI(JSON.stringify(fileToKey));
             }
@@ -127,9 +128,9 @@ export default class S3Album extends Component {
     }
 
     componentDidUpdate(prevProps, prevState) {
-        if (this.props.path == prevProps.path &&
-            this.props.ts == prevProps.ts &&
-            this.props.select == prevProps.select
+        if (this.props.path === prevProps.path &&
+            this.props.ts === prevProps.ts &&
+            this.props.select === prevProps.select
         ) {
             return;
         }
@@ -174,9 +175,9 @@ export default class S3Album extends Component {
             if (!item.contentType) { item.contentType = this.contentType(item); }
         });
 
-        list = this.filter(list);
-        list = this.sort(list);
-        this.setState({ items: list });
+        let items = this.filter(list);
+        items = this.sort(items);
+        this.setState({ items });
     }
 
     filter(list) {
@@ -218,7 +219,7 @@ export default class S3Album extends Component {
 
         const list = items.map(item => {
             const isText = item.contentType && JS.isTextFile(item.contentType);
-            return isText? <S3Text
+            return isText ? <S3Text
                              key={item.key}
                              textKey={item.key}
                              theme={theme}
@@ -237,7 +238,7 @@ export default class S3Album extends Component {
                              translate={translateItem}
                              level={level}
                              onClick={() => this.handleClick(item)}
-                           />
+                           />;
         });
         return (
             <div>
@@ -254,6 +255,6 @@ export default class S3Album extends Component {
                         : null
                 }
             </div>
-        )
+        );
     }
 }