+++ packages/aws-amplify-react/src/Widget/Picker.js
@@ -11,7 +11,8 @@
  * and limitations under the License.
  */
 
-import React, { Component } from 'react';
+import * as React from 'react';
+import { Component } from 'react';
 
 import { I18n, ConsoleLogger as Logger } from '@aws-amplify/core';
 import AmplifyTheme from '../Amplify-UI/Amplify-UI-Theme';
@@ -19,7 +20,7 @@ import { Button, PhotoPickerButton } from '../Amplify-UI/Amplify-UI-Components-R
 
 const PickerPicker = {
     position: 'relative'
-}
+};
 
 // const PickerButton = {
 //     width: '10em',
@@ -37,13 +38,13 @@ const PickerInput = {
     top: 0,
     opacity: 0,
     cursor: 'pointer'
-}
+};
 
 const logger = new Logger('Picker');
 
 export default class Picker extends Component {
     handleInput(e) {
-        var that = this;
+        const that = this;
 
         const file = e.target.files[0];
         if (!file) { return; }
@@ -54,10 +55,10 @@ export default class Picker extends Component {
         const { onPick } = this.props;
         if (onPick) {
             onPick({
-                file: file,
-                name: name,
-                size: size,
-                type: type
+                file,
+                name,
+                size,
+                type
             });
         }
     }
@@ -88,6 +89,6 @@ export default class Picker extends Component {
                     onChange={(e) => this.handleInput(e)}
                 />
             </div>
-        )
+        );
     }
 }